remove test polylines and move osmdata to services
