+++ lib/components/hikingmap.dart
@@ -5,7 +5,7 @@ import 'package:latlong/latlong.dart';
 import 'package:location/location.dart';
 import 'package:flutter/services.dart';
 import 'package:hiking4nerds/components/types.dart';
-import 'package:hiking4nerds/osmdata.dart';
+import 'package:hiking4nerds/services/osmdata.dart';
 
 class HikingMap extends StatefulWidget {
   @override
@@ -118,34 +118,9 @@ class HikingMapState extends State<HikingMap> {
   }
 
   PolylineLayerOptions getPolyLineLayerOptions() {
-    var points = <LatLng>[
-      LatLng(52.5, 13.455),
-      LatLng(52.5, 13.46),
-      LatLng(52.5, 13.47),
-      LatLng(52.52, 13.48),
-      LatLng(52.53, 13.49),
-      LatLng(52.53, 13.48),
-      LatLng(52.57, 13.5),
-      LatLng(52.58, 13.5),
-      LatLng(52.59, 13.51),
-      LatLng(52.5, 13.5),
-      LatLng(52.5, 13.455),
-    ];
-
-    var points2 = <LatLng>[
-      LatLng(52.5, 13.455),
-      LatLng(52.53, 13.458),
-      LatLng(52.54, 13.459),
-      LatLng(52.58, 13.459),
-      LatLng(52.58, 13.5),
-      LatLng(52.7, 13.55),
-    ];
-
 
     var polylineLayerOptions = new PolylineLayerOptions(
       polylines: [
-        Polyline(points: points, strokeWidth: 4.0, color: Colors.purple),
-        Polyline(points: points2, strokeWidth: 4.0, color: Colors.green),
       ],
     );
